PRC-56 : Add assertions to create contact integration tests
Assert that the create contact API is called with the correct parameters in the integration tests. While doing this I discovered a bug in that we were sending a title and middle name with an empty string instead of null.

Also added some missing tests to check that you can't just enter whitespace for the name.