PRC-57: Support edit DOB from check answers (#37)
Discovered several issues with the old data model. For one `Date` is converted back to string when it's persisted in the session. No idea how this worked so well up until now but if you need to parse it back it's a string. Instead, keep the originally entered day, month and year fields to construct a `Date` as needed.

Also, javascript was doing javascript things with the values for the isDobKnown field being true and false. Converted them to be `'Yes' | 'No'` type instead which matches the actual answers.