PRC-122: Improvements to cypress page chaining (#53)
Added support for `verifyOnPage` where the page takes a constructor argument, such as contact name.

Added new `ContinuablePage` base class for pages with a continue button. This adds a function for continuing to a specific page, verifying you are on the page and updating the chain to the new page type, e.g., `continueTo(SelectNextOfKinPage, 'Last, First')`

Restructured the create journey tests to reduce duplication and leverage the new features.