PRC-188: Reset journey on switch mode (#83)
When switching from add existing to add new it was prefilling the name and date of birth. This could happen when you select "No" on contact confirmation and then click "The contact is not listed".

You could also proceed a new journey and relationship details would be persisted if you then went back to search results and chose existing contact.

Now the journey is reset whenever you select mode (click a result or not listed).