PRC-163: Address fixes (#174)
Could not link a contact if they had an address with no address type.

Mail address question had wrong label.

Postcode was missing in several places addresses are rendered.

There was no way to add an address on a new contact as you only get the link to view all addresses if you have an address.  PRC-163 defined an add address link so adding it here.