PRC-450: Add missing legends (#240)
* PRC-450: Add missing legends

Questions with radio button answers that are also the H1 did not have an accessible legend. Moved the H1 to the fieldset for these questions as per GDS guidelines. Annoyingly, fieldset does not allow setting custom attributes so accessors in tests changed to class.

* PRC-450: Add check for radio group legend

Radio groups should be in a fieldset with a legend for accessibility reasons. This applies to both cases when a radio group is only 1 input on a page as well as if the radio group is the only input on the page and the legends becomes the h1.