PRC-454: Add / change date of death
Show a new link on main contact details tab to record a date of death if the contact does not have a deceased date. Link is also show on edit contact details.

Date of death can only be changed from edit contact details.

New combined date error messages required a minor change to validation to exclude empty error messages. This allows the all fields to be highlighted but produced an empty link which was flagged by accessibility checks.