PRC-460: Add identity documents (#307)
Support for adding multiple identity documents at once. Initial implementation of add-another pattern.

Re-wrote the deduplication of error messages to support nested paths.

Error links are all correct and give focus to the correct field.

Complication in the controller to support different actions is only required when javascript is not enabled otherwise the adding/removing of elements happens inline.