PRC-615: Add audit event details for security (#405)
For potentially security related modifications we now include the details of the action in the audit event. This includes anywhere that sets isApprovedVisitor or adds or changes a restriction.

This would allow the audit team to see who approved or unapproved a visitor or set restrictions even if they have since been modified again.