PRC-619: RBAC enforced on visit approval
Apply RBAC to approve to visit pages so you will be redirected to not found if you attempt to GET or POST on a relevant controller.

Hide change visit approval link if you do not have permission and offer a hint as to why.

Remove visit approval from the create & link contact journeys if you do not have permission.