PRC-703: Edit relationship dupe screen (#455)
* PRC-703: Edit relationship dupe screen

When editing the relationship type, if we receive a conflict from the API then show a new screen which shows the existing relationships between a prisoner and contact with an option to go to the duplicate record or contact list.

* PRC-703: Refactor edit relationship to prisoner

Now instead of being a bounce page it re-uses the mini-journey for edit contact relationship type which includes the duplicate handling so now both methods handle duplicates properly.